+++ postgres/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.3
+
++ Added `parseValue` to `Adapter`
+
 ## 2.1.1
 
 + Dart2 support