+++ query/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.2
+
++ Added `parseValue` to adapter
+
 ## 2.1.1
 
 + Dart2 support