+++ sqflite/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.4
+
++ Added `parseValue` to parse values from Database
+
 ## 2.1.2
 
 + sqlite-ify create statement